Fix CORS w/basic auth
After updating Opera, it didn't work anymore.
It seems I was doing it a bit wrong.
On the client side, force withCredentials, so that the basic auth data is sent along, while on the server side,
Access-Control-Allow-Credentials must be sent within the headers.

Also updated versions for jQuery and Backbone.